+++ COPS/CHANGELOG.md
@@ -1,6 +1,12 @@
 # HA COPS Changelog
 
-## [1.9.2] - 2023-07-12
+## [1.9.4] - 2023-07-13
+
+- Bug fix so address to send books by mail to is displayed on customisation page
+- Limit sending books to a single email address
+- Check that email address is valid before sending
+
+## [1.9.3] - 2023-07-12
 
 - Prefer EPUB but allow sending MOBI, AZW3 or PDF in absence of EPUB